C++ programs!!
If someone who is checking this codes out , it contains all the different types of functions, syntax that can be used in a C++ program. Also keep in mind that for graphics program it is preferred to use Turbo C++ or dev C++ and don't forget to include graphics.h in programs related to graphics programs.